+++ test/meta_tests.jl
@@ -15,7 +15,7 @@
         "oracle_tests.jl" => 4,
         "polish_tests.jl" => 3,
         "scaling_tests.jl" => 7,
-        "setup_tests.jl" => 10,
+        "setup_tests.jl" => 11,
         "solve_tests.jl" => 21,
         "trim_tests.jl" => 1,
         "update_tests.jl" => 5,